Move tree-view filter to separated component
