+++ src/ggrc-client/js/components/tree/tests/tree-widget-container_spec.js
@@ -12,7 +12,6 @@ import {
 } from '../../../../js_specs/spec_helpers';
 import * as TreeViewUtils from '../../../plugins/utils/tree-view-utils';
 import * as WidgetsUtils from '../../../plugins/utils/widgets-utils';
-import * as AdvancedSearch from '../../../plugins/utils/advanced-search-utils';
 import * as NotifierUtils from '../../../plugins/utils/notifiers-utils';
 import * as MegaObjectUtils from '../../../plugins/utils/mega-object-utils';
 import tracker from '../../../tracker';
@@ -93,6 +92,7 @@ describe('tree-widget-container component', function () {
         filter,
         request,
       });
+      vm.attr('currentFilter', filter);
 
       loadItems = vm.loadItems.bind(vm);
       spyOn(tracker, 'start').and.returnValue(() => {});
@@ -199,246 +199,6 @@ describe('tree-widget-container component', function () {
 
       _widgetShown();
       expect(vm.loadItems).toHaveBeenCalled();
-    });
-  });
-
-  describe('openAdvancedFilter() method', function () {
-    it('copies applied filter and mapping items', function () {
-      let appliedFilterItems = new canList([
-        AdvancedSearch.create.attribute(),
-      ]);
-      let appliedMappingItems = new canList([
-        AdvancedSearch.create.mappingCriteria({
-          filter: AdvancedSearch.create.attribute(),
-        }),
-      ]);
-      vm.attr('advancedSearch.appliedFilterItems', appliedFilterItems);
-      vm.attr('advancedSearch.appliedMappingItems', appliedMappingItems);
-      vm.attr('advancedSearch.filterItems', canList());
-      vm.attr('advancedSearch.mappingItems', canList());
-
-      vm.openAdvancedFilter();
-
-      expect(vm.attr('advancedSearch.filterItems').attr())
-        .toEqual(appliedFilterItems.attr());
-      expect(vm.attr('advancedSearch.mappingItems').attr())
-        .toEqual(appliedMappingItems.attr());
-    });
-
-    it('opens modal window', function () {
-      vm.attr('advancedSearch.open', false);
-
-      vm.openAdvancedFilter();
-
-      expect(vm.attr('advancedSearch.open')).toBe(true);
-    });
-
-    it('should add "parentInstance" when isObjectContextPage is TRUE' +
-    'and "parentInstance" is empty', () => {
-      const parentInstance = {id: 1, type: 'Audit'};
-      spyOn(CurrentPageUtils, 'isObjectContextPage').and.returnValue(true);
-      spyOn(AdvancedSearch.create, 'parentInstance')
-        .and.returnValue({type: 'parentInstance', value: parentInstance});
-
-      vm.attr('advancedSearch.parentInstance', null);
-
-      vm.openAdvancedFilter();
-
-      expect(AdvancedSearch.create.parentInstance).toHaveBeenCalled();
-      expect(vm.attr('advancedSearch.parentInstance.value').serialize())
-        .toEqual(parentInstance);
-    });
-
-    it('should NOT add "parentInstance" because it already set', () => {
-      spyOn(CurrentPageUtils, 'isObjectContextPage').and.returnValue(true);
-      spyOn(AdvancedSearch.create, 'parentInstance');
-
-      vm.attr('advancedSearch.parentInstance', {id: 1});
-
-      vm.openAdvancedFilter();
-
-      expect(AdvancedSearch.create.parentInstance).not.toHaveBeenCalled();
-      expect(vm.attr('advancedSearch.parentInstance')).not.toBeNull();
-    });
-
-    it('should NOT add "parentInstance" because isObjectContextPage is FALSE',
-      () => {
-        spyOn(CurrentPageUtils, 'isObjectContextPage').and.returnValue(false);
-        spyOn(AdvancedSearch.create, 'parentInstance');
-
-        vm.attr('advancedSearch.parentInstance', null);
-
-        vm.openAdvancedFilter();
-
-        expect(AdvancedSearch.create.parentInstance).not.toHaveBeenCalled();
-        expect(vm.attr('advancedSearch.parentInstance')).toBeNull();
-      }
-    );
-
-    it('should filter parentItems and exclude parentInstance', () => {
-      const parentInstance = {id: 1, type: 'Audit'};
-      spyOn(CurrentPageUtils, 'isObjectContextPage').and.returnValue(true);
-      spyOn(AdvancedSearch.create, 'parentInstance')
-        .and.returnValue({type: 'parentInstance', value: parentInstance});
-
-      vm.attr('advancedSearch.parentInstance', null);
-
-      vm.attr('advancedSearch.appliedParentItems', [
-        {value: {id: 1, type: 'Audit'}},
-        {value: {id: 2, type: 'Audit'}},
-        {value: {id: 1, type: 'Program'}},
-      ]);
-
-      vm.openAdvancedFilter();
-
-      expect(vm.attr('advancedSearch.parentItems').serialize()).toEqual([
-        {value: {id: 2, type: 'Audit'}},
-        {value: {id: 1, type: 'Program'}},
-      ]);
-
-      expect(vm.attr('advancedSearch.parentInstance.value').serialize())
-        .toEqual({
-          id: 1,
-          type: 'Audit',
-        });
-    });
-  });
-
-  describe('applyAdvancedFilters() method', function () {
-    let filterItems = new canList([
-      AdvancedSearch.create.attribute(),
-    ]);
-    let mappingItems = new canList([
-      AdvancedSearch.create.mappingCriteria({
-        filter: AdvancedSearch.create.attribute(),
-      }),
-    ]);
-    beforeEach(function () {
-      vm.attr('advancedSearch.filterItems', filterItems);
-      vm.attr('advancedSearch.mappingItems', mappingItems);
-      vm.attr('advancedSearch.appliedFilterItems', canList());
-      vm.attr('advancedSearch.appliedMappingItems', canList());
-      spyOn(vm, 'onFilter');
-      spyOn(AdvancedSearch, 'buildFilter')
-        .and.callFake(function (items, request) {
-          request.push({name: 'item'});
-        });
-      spyOn(QueryParser, 'joinQueries');
-    });
-
-    it('copies filter and mapping items to applied', function () {
-      vm.applyAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.appliedFilterItems').attr())
-        .toEqual(filterItems.attr());
-      expect(vm.attr('advancedSearch.appliedMappingItems').attr())
-        .toEqual(mappingItems.attr());
-    });
-
-    it('initializes advancedSearch.filter property', function () {
-      QueryParser.joinQueries.and.returnValue({
-        name: 'test',
-      });
-      vm.attr('advancedSearch.filter', null);
-
-      vm.applyAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.filter.name')).toBe('test');
-    });
-
-    it('initializes advancedSearch.request property', function () {
-      vm.attr('advancedSearch.request', canList());
-
-
-      vm.applyAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.request.length')).toBe(3);
-    });
-
-    it('closes modal window', function () {
-      vm.attr('advancedSearch.open', true);
-
-      vm.applyAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.open')).toBe(false);
-    });
-
-    it('calls onFilter() method', function () {
-      vm.applyAdvancedFilters();
-
-      expect(vm.onFilter).toHaveBeenCalled();
-    });
-  });
-
-  describe('removeAdvancedFilters() method', function () {
-    beforeEach(function () {
-      spyOn(vm, 'onFilter');
-    });
-
-    it('removes applied filter and mapping items', function () {
-      vm.attr('advancedSearch.appliedFilterItems', new canList([
-        {title: 'item'},
-      ]));
-      vm.attr('advancedSearch.appliedMappingItems', new canList([
-        {title: 'item'},
-      ]));
-
-      vm.removeAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.appliedFilterItems.length')).toBe(0);
-      expect(vm.attr('advancedSearch.appliedMappingItems.length')).toBe(0);
-    });
-
-    it('cleans advancedSearch.filter property', function () {
-      vm.attr('advancedSearch.filter', {});
-
-      vm.removeAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.filter')).toBe(null);
-    });
-
-    it('closes modal window', function () {
-      vm.attr('advancedSearch.open', true);
-
-      vm.removeAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.open')).toBe(false);
-    });
-
-    it('calls onFilter() method', function () {
-      vm.removeAdvancedFilters();
-
-      expect(vm.onFilter).toHaveBeenCalled();
-    });
-
-    it('resets advancedSearch.request list', function () {
-      vm.attr('advancedSearch.request', new canList([{data: 'test'}]));
-
-      vm.removeAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.request.length')).toBe(0);
-    });
-  });
-
-  describe('resetAdvancedFilters() method', function () {
-    it('resets filter items', function () {
-      vm.attr('advancedSearch.filterItems', new canList([
-        {title: 'item'},
-      ]));
-
-      vm.resetAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.filterItems.length')).toBe(0);
-    });
-
-    it('resets mapping items', function () {
-      vm.attr('advancedSearch.mappingItems', new canList([
-        {title: 'item'},
-      ]));
-
-      vm.resetAdvancedFilters();
-
-      expect(vm.attr('advancedSearch.mappingItems.length')).toBe(0);
     });
   });
 
@@ -745,6 +505,7 @@ describe('tree-widget-container component', function () {
         filter,
         request,
       });
+      vm.attr('currentFilter', filter);
     });
 
     it('starts export correctly', () => {
@@ -761,51 +522,6 @@ describe('tree-widget-container component', function () {
         'info',
         exportMessage,
         {data: true});
-    });
-  });
-
-  describe('applySavedSearch() method', () => {
-    let method;
-
-    beforeEach(() => {
-      spyOn(vm, 'clearAppliedSavedSearch');
-      vm.attr('filterIsDirty', false);
-      vm.attr('savedSearchPermalink', '');
-
-      method = vm.applySavedSearch.bind(vm);
-    });
-
-    it('should call "clearAppliedSavedSearch" when selectedSavedSearch is null',
-      () => {
-        method(null);
-        expect(vm.clearAppliedSavedSearch).toHaveBeenCalled();
-      }
-    );
-
-    it('should call "clearAppliedSavedSearch" when filter is dirty', () => {
-      vm.attr('filterIsDirty', true);
-
-      method({});
-      expect(vm.clearAppliedSavedSearch).toHaveBeenCalled();
-    });
-
-    it('should set permalink when selectedSavedSearch is not empty ' +
-    'and filter is NOT dirty', () => {
-      spyOn(AdvancedSearch, 'buildSearchPermalink').and.returnValue('link');
-
-      method(new canMap({id: 5}));
-      expect(AdvancedSearch.buildSearchPermalink).toHaveBeenCalled();
-      expect(vm.attr('savedSearchPermalink')).toEqual('link');
-    });
-
-    it('should set appliedSavedSearch when selectedSavedSearch is not empty ' +
-    'and filter is NOT dirty', () => {
-      const selectedSavedSearch = {id: 123};
-      spyOn(AdvancedSearch, 'buildSearchPermalink').and.returnValue('link');
-
-      method(new canMap(selectedSavedSearch));
-      expect(vm.attr('appliedSavedSearch').serialize())
-        .toEqual(selectedSavedSearch);
     });
   });
 
@@ -958,155 +674,5 @@ describe('tree-widget-container component', function () {
       vm.addServiceColumns(columns);
       expect(columns).toEqual(expectedOutput);
     });
-  });
-});
-
-describe('loadSavedSearch() function', () => {
-  const viewModel = new canMap({
-    router: {
-      saved_search: 1,
-    },
-    loading: false,
-    modelName: 'Control',
-    advancedSearch: {},
-    removeAdvancedFilters: () => {},
-    applyAdvancedFilters: () => {},
-  });
-
-  let treeFunction;
-
-  beforeAll(() => {
-    treeFunction = loadSavedSearch;
-  });
-
-  it('should set "loading" flag to true while loading', (done) => {
-    let dfd = $.Deferred();
-
-    spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-
-    let loadDfd = treeFunction(viewModel);
-    expect(viewModel.attr('loading')).toBeTruthy();
-
-    loadDfd.then(() => {
-      expect(viewModel.attr('loading')).toBeFalsy();
-      done();
-    });
-
-    dfd.resolve({});
-  });
-
-  it('should call "removeAdvancedFilters" when search response is null',
-    (done) => {
-      let dfd = $.Deferred();
-
-      spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-      spyOn(viewModel, 'removeAdvancedFilters');
-
-      treeFunction(viewModel).then(() => {
-        expect(viewModel.removeAdvancedFilters).toHaveBeenCalled();
-        done();
-      });
-
-      dfd.resolve({SavedSearch: null});
-    }
-  );
-
-  it('should call "removeAdvancedFilters" when search is NOT AdvancedSearch',
-    (done) => {
-      let dfd = $.Deferred();
-
-      spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-      spyOn(viewModel, 'removeAdvancedFilters');
-
-      treeFunction(viewModel).then(() => {
-        expect(viewModel.removeAdvancedFilters).toHaveBeenCalled();
-        done();
-      });
-
-      dfd.resolve({SavedSearch: {
-        search_type: 'GlobalSearch',
-        object_type: 'Control',
-      }});
-    }
-  );
-
-  it('should call "removeAdvancedFilters" when search.object_type' +
-  ' is NOT equal to viewModel.modelName',
-  (done) => {
-    let dfd = $.Deferred();
-
-    spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-    spyOn(viewModel, 'removeAdvancedFilters');
-
-    treeFunction(viewModel).then(() => {
-      expect(viewModel.removeAdvancedFilters).toHaveBeenCalled();
-      done();
-    });
-
-    dfd.resolve({SavedSearch: {
-      search_type: 'AdvancedSearch',
-      object_type: 'Regulation',
-    }});
-  });
-
-  it('should call "removeAdvancedFilters" when search loading was failed',
-    (done) => {
-      let dfd = $.Deferred();
-
-      spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-      spyOn(viewModel, 'removeAdvancedFilters');
-
-      treeFunction(viewModel).fail(() => {
-        expect(viewModel.removeAdvancedFilters).toHaveBeenCalled();
-        done();
-      });
-
-      dfd.reject();
-    }
-  );
-
-  it('should call "applyAdvancedFilters" when search appropriate to viewModel',
-    (done) => {
-      let dfd = $.Deferred();
-
-      spyOn(SavedSearch, 'findOne').and.returnValue(dfd);
-      spyOn(viewModel, 'applyAdvancedFilters');
-      spyOn(AdvancedSearch, 'parseFilterJson');
-
-      treeFunction(viewModel).then(() => {
-        expect(viewModel.applyAdvancedFilters).toHaveBeenCalled();
-        done();
-      });
-
-      dfd.resolve({SavedSearch: {
-        search_type: 'AdvancedSearch',
-        object_type: 'Control',
-      }});
-    }
-  );
-});
-
-describe('filterParentItems() method', () => {
-  it('should exclude parentInstance from parentItem', () => {
-    const parentInstance = {
-      value: {
-        id: 5,
-        type: 'Control',
-      },
-    };
-
-    const parentItems = [
-      {value: {id: 7, type: 'Regulation'}},
-      {value: {id: 5, type: 'Regulation'}},
-      {value: {id: 5, type: 'Control'}},
-      {value: {id: 6, type: 'Control'}},
-    ];
-
-    const result = filterParentItems(parentInstance, parentItems);
-    expect(result).toEqual([
-      {value: {id: 7, type: 'Regulation'}},
-      {value: {id: 5, type: 'Regulation'}},
-      {value: {id: 6, type: 'Control'}},
-    ]);
   });
 });
