+++ src/ggrc-client/js/plugins/utils/tree-view-utils.js
@@ -49,6 +49,7 @@ import {
   setTreeViewHeaders,
 } from './display-prefs-utils';
 import TreeViewConfig from '../../apps/base_widgets';
+import QueryParser from '../../generated/ggrc_filter_query_parser';
 
 /**
 * TreeView-specific utils.
@@ -702,6 +703,27 @@ function startExport(
   });
 }
 
+/**
+ * Concatenation active filters.
+ *
+ * @param {String} filter - Parsed filter string
+ * @param {Object} options - Filter parameters
+ * @return {string} - Result of concatenation filters.
+ * @private
+ */
+function concatFilters(filter, options) {
+  if (filter) {
+    filter = QueryParser.joinQueries(
+      filter,
+      options.query.attr(),
+      'AND');
+  } else if (options.query) {
+    filter = options.query;
+  }
+
+  return filter;
+}
+
 export {
   getAvailableAttributes,
   getColumnsForModel,
@@ -714,4 +736,5 @@ export {
   getVisibleColumnsConfig,
   isDirectlyRelated,
   startExport,
+  concatFilters,
 };
